Posts page rendering and working correctly
