--- instagram/src/App.css
@@ -1,10 +1,0 @@
-.App {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.posts {
-  max-width: 600px;
-  width: 100%;
-}
